+++ mutwo/core_version/__init__.py
@@ -1,4 +1,4 @@
-MAJOR, MINOR, PATCH = 0, 62, 0
+MAJOR, MINOR, PATCH = 0, 63, 0
 
 VERSION = f"{MAJOR}.{MINOR}.{PATCH}"
 """The version of the package ``mutwo.core``."""