+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - `cut_out` method of `Envelope`
 - `split_at` method of `Envelope`
+- `TempoEnvelope.convert` corner case issue
 
 
 ## [0.62.0] - 2022-10-06