+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.63.0] - 2022-11-03
+
 ### Added
 - `EmptyEnvelopeError` in case user tries to call `value_at` on an envelope without any events.
 - `split_at` method for `Envelope`